LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-onsignalingstatechanged.https-expected.txt


PASS Negotiation methods fire signalingstatechange events
PASS Closing a PeerConnection should not fire signalingstatechange event
PASS signalingstatechange is the first event to fire
